# docs/openapi.yaml
openapi: 3.0.3
info:
  title: iag-traceability API
  version: 1.0.0
  description: Custody events, chain read model, and public QR for IAG coffee traceability.
servers:
  - url: /api/v1/traceability
paths:
  /api/v1/batches/{businessId}/chain:
    get:
      summary: Batch chain of custody
      security:
        - bearerAuth: []
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chain nodes from trace_events
  /api/v1/events:
    get:
      summary: List custody events for an entity
      security:
        - bearerAuth: []
      parameters:
        - name: entity_type
          in: query
          required: true
          schema:
            type: string
        - name: entity_business_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event list
    post:
      summary: Append a custody event
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Event created
  /api/v1/lots/{businessId}/publish:
    post:
      summary: Publish public QR for export lot
      security:
        - bearerAuth: []
      responses:
        "200":
          description: QR token and URL
  /api/v1/lots/{businessId}/revoke:
    post:
      summary: Revoke public QR for export lot
      security:
        - bearerAuth: []
      responses:
        "200":
          description: QR revoked
  /api/v1/admin/audit-logs:
    get:
      summary: List HTTP API audit entries
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Paginated audit log rows
  /api/v1/admin/monitoring/summary:
    get:
      summary: Admin monitoring summary (24h API stats + domain counts)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Monitoring summary
  /api/v1/admin/monitoring/activity:
    get:
      summary: Recent HTTP API activity
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent audit rows
  /api/v1/admin/monitoring/dead-letters:
    get:
      summary: Recent dead-lettered (unmapped) Kafka events
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Recent dead-letter rows (event_id, topic, event_type, reason, received_at)
  /public/q/{token}:
    get:
      summary: Public coffee story (QR payload v2)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public story JSON (Cache-Control public)
  /public/q/{token}/qr.png:
    get:
      summary: Public QR code image (PNG) encoding the story URL
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PNG image of the QR code
          content:
            image/png:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT